fluid: Improve the code, and benchmark both Futhark, C, and Accelerate.
The Futhark code now has less recursive calls, hopefully resulting in less of a
code explosion.

Also add `unsafe` to all index accesses to make compiling to OpenCL possible.

Extend the C source to have both `fluid-benchmark` and
`fluid-visualize-densities` programs with the same purpose as the Futhark
programs of the same name, to make it possible to either measure the raw speed
of working on a grid of some size over n steps for benchmarking, or to save all
density frames for visualization.

Make it possible to visualize with a predetermined seed.  This is useful for
comparing C output frames and Futhark output frames.

Figure out how to measure Accelerate on the same input dimensions as the C and
Futhark versions, to make the results compareable (see the `benchmark`
directory).  Its backend defaults to CUDA on Napoleon, so that's nice.

Probably some other things as well.


Former-commit-id: 792021fa398a4833e2d3c6814e470e55f0fb60c1