Change the fluid benchmark to default to doing a single step.
I confused Accelerate's "simulationSteps" configuration variable to mean the
number of steps in the simulation, but Accelerate always runs only one step.
"simulationSteps" instead means the number of runs of their linear solver...
This was previously fixed to the constant 20 in the Futhark version, based on
the C code, but can now be adjusted like the other parameters.  Accelerate
defaults to 40, so that's what we're defaulting to as well.

This explains a lot.


Former-commit-id: 81a4fb162fb8e6f7712638d8d27b9d5f73e77a32