Add larger datasets for some benchmarks
Some of the datasets in our benchmark suite are way too small for modern GPUs
like the A100. I've tried to add a few new ones where it felt appropriate.

Some benchmarks that still have too small datasets are bfast (not cloudy) and
hashcat, but it seems like adding those is not as trivial.