Use get-data.sh to manage datasets from ERDA
This removes the dependency on git-annex and greatly simplifies that management
of external data.

The idea is to maintain an external-data.txt file which contains lines of the
form:

OUTPUT URL SHA

where OUTPUT is the path of the file to manage with get-data.sh, preferably
something in .external-data, which has been added to .gitignore. URL is the
url to the data hosted on ERDA and the SHA is the SHA256 checksum of the file.

After checking out the repository, simply run ./get-data.sh external-data.txt
to download the data dependencies.