It is fine to return gigantic things, actually.
This makes the benchmarks more similar to what they are actually
supposed to compute.  And that bounds-checked stuff at the end can get
surprisingly slow!

I suspect the old change was due to (now-removed) backends that had a
hard time printing large values efficiently, but this is not
observable now.  I'm not sure 'futhark bench' even asks for the
results when there is no validation.