Use NW from futhark-mem-sc22
This version of NW is faster, even though short-circuiting doesn't currently
apply. I'm not sure why.

The downside is that the new version doesn't support the smallest datasets, but
they are so small as to be insignificant anyway.

```console
$ ~/src/futhark/tools/cmp-bench-json.py {old,new}.json
In old.json but not new.json: program rodinia/nw/nw.fut dataset data/small.in
In old.json but not new.json: program rodinia/nw/nw.fut dataset data/tiny.in

rodinia/nw/nw.fut
  data/large.in:                                                        1.15x (mem: 1.47x@device)
  data/medium.in:                                                       1.32x (mem: 1.45x@device)
```